Preparing use of libkface in a multithreaded environment.
Current threading guidelines of libface:
FaceDetect is fully reentrant, only needs a good amount of memory per object 
for the cascades. That means there could be two concurrent objects in use.
The Eigenfaces, in principle, is also reentrant, but as soon as training is
done, it becomes desirable to only have one concurrent object for recognition
+ training because, obviously, the training data is altered in the object that
is trained.

=>

Add two wrapper classes to reflect these properties. Both
- offer automatic deferred creation - the backend is created only at first real use
- only init a core for their task, that is, either Detection or Recognition

FaceDetector is specified as reentrant but not thread safe.

RecognitionDatabase controls its creation, limiting to one object per config path.
By using an internal mutex, this object is thread-safe.
(Note: There are some caveats with reference counted Priv classes and storing
a hash to existing objects. The relevant code needs testing.)
 
Both classes are built around Database without own functionality, 
so it is for the simple usecase completely sufficient to use a Database object.
For the multithreaded purpose the wrapper classes provide additional semantics.



svn path=/branches/extragear/graphics/libkface/; revision=1171026