look also for OpenCV XML cascades
In addition to the KDE directories, look also in the OpenCV data paths;
this way it is possible to skip the libkface-provided cascade XML files
and just use the OpenCV-provided ones (could help having cascade data
in sync with the installed OpenCV).

A bit hackish (using xdgdata-apps and then going up one level), but
since KStandardDirs does not allow just looking in $XDG_DATA_DIRS,
this is the best solution I've found so far, and should for it
(for KDE 4).